Refactor input components.
